Use C++17 instead of C++20 by default
This commit removes the dependence on `std::span` which means that this
header can be used with C++17. Various methods that took a `std::span`
now have overloads that take a pointer/length as well as a `const
vector<T>&`. Methods that give a `std::span` now have separate accessor
methods for the pointer and the length.

The hardest one to migrate was the arguments provided to function
closures, and for that purpose a local `Span` class was created.

In all cases support for C++20 is in theory detected at compile time and
`std::span` support is enabled (methods are added back, etc). The
`wasmtime::Span` class should be auto-convertible to `std::span` in
theory (if I understand C++ correctly).