Adjusting the function passed to func_new/func_wrap to be universal reference so that we can perfectly forward it on.
The issue, as shown by the test cases, is that previously the function was always copied when passed to func_new/wrap. This works fine for pointers, but when using a callable it's unnecessarily copied. Furthermore, for callables which aren't allowed to be copied (like lambda's with refernces, compilation actually fails.

The fix is to move the callable when the callable is passed in as an xvalue.